+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.40]
+### Added
+- `StyledToggle` component
+- `noPadding` prop in `DynamicTable` to conditionally add `no-padding` class
+
+### Refactored
+- Export `ButtonAppearance` from `Button` and `PopoverPlacement` from `InlinePopover`
+- Move some packages that aren't used after build to devDependencies
+- Set `label` prop in `Checkbox` and `StyledRadioTabVertical` as optional
+
 ## [0.1.39]
 ### Fixed
 - Build is not up to date with src